+++ agave/models/__init__.py
@@ -1,0 +1,3 @@
+__all__ = ['BaseModel']
+
+from .base import BaseModel
