+++ agave/models/base.py
@@ -1,7 +1,6 @@
-from collections import OrderedDict
 from typing import ClassVar, Dict
 
-from .helpers import DynamicLazyReferenceField, mongo_to_dict
+from ..lib.mongoengine.model_helpers import mongo_to_dict
 
 
 class BaseModel:
@@ -9,23 +8,7 @@ class BaseModel:
     _hidden: ClassVar = []
 
     def __init__(self, *args, **values):
-        values = self.priority_dynamic_field(values)
         return super().__init__(*args, **values)
-
-    def priority_dynamic_field(self, values):
-        dynamic_fields = [
-            name
-            for name, field in self._fields.items()
-            if isinstance(field, DynamicLazyReferenceField)
-        ]
-        if not dynamic_fields:
-            return values
-        for field_name in dynamic_fields:
-            if field_name not in values:
-                continue
-            values = OrderedDict(values)
-            values.move_to_end(field_name)
-        return values
 
     def to_dict(self) -> Dict:
         private_fields = [f for f in dir(self) if f.startswith('_')]
