+++ examples/chalicelib/models/transactions.py
@@ -1,6 +1,6 @@
 from mongoengine import Document, FloatField, StringField
 
-from agave.models.base import BaseModel
+from agave.models import BaseModel
 from agave.models.helpers import uuid_field
