+++ requirements-test.txt
@@ -1,4 +1,5 @@
 pytest==6.0.*
+pytest-freezegun==0.4.*
 pytest-cov==2.10.*
 black==20.8b1
 isort==5.5.*
