+++ setup.cfg
@@ -2,7 +2,7 @@
 test=pytest
 
 [tool:pytest]
-addopts = -p no:warnings -v --cov-report term-missing --cov=agave
+addopts = -p no:warnings -v --cov-report html --cov=agave
 
 [flake8]
 inline-quotes = '
@@ -18,7 +18,7 @@ ignore_missing_imports = true
 [coverage:run]
 source = agave
 branch = True
-omit = tests/*,venv/*,agave/models/helpers.py,agave/models/base.py
+omit = tests/*,venv/*,agave/lib/mongoengine/enum_field.py,agave/lib/mongoengine/model_helpers.py
 
 [coverage:report]
 precision = 2
